rename paragraph tests to _golden_test.dart
