+++ lib/web_ui/test/html/paragraph/text_placeholders_golden_test.dart
